git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@11565 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/REPLICA/fix_event_prd.cpp
@@ -91,8 +91,8 @@ void FixEventPRD::restart(char *buf)
   double *list = (double *) buf;
 
   event_number = static_cast<int> (list[n++]);
-  event_timestep = static_cast<int> (list[n++]);
-  clock = static_cast<int> (list[n++]);
+  event_timestep = static_cast<bigint> (list[n++]);
+  clock = static_cast<bigint> (list[n++]);
   replica_number = static_cast<int> (list[n++]);
   correlated_event = static_cast<int> (list[n++]);
   ncoincident = static_cast<int> (list[n++]);
